One to many mapping on organisation and partnerorganisation (also the same for users and files in previous commit

+++ src/main/java/graphium/graphiumteam8/repository/OrganisationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface OrganisationRepository extends JpaRepository<Organisation, Long> {
+public interface OrganisationRepository extends JpaRepository<Organisation, Integer> {
 
     List<Organisation> findAll();
